sports-highlights-pipeline/app/requirements.txt
boto3>=1.26
requests>=2.28
python-dotenv>=1.0
